Package challenge renamed to store
