+++ store/src/main/java/br/com/estudo/storechallenge/store/controller/PaymentController.java
@@ -1,6 +1,6 @@
-package br.com.estudo.storechallenge.challenge.controller;
+package br.com.estudo.storechallenge.store.controller;
 
-import br.com.estudo.storechallenge.challenge.service.PaymentService;
+import br.com.estudo.storechallenge.store.service.PaymentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RestController;
